Followed directions.... :)
Day2 HTML update was the result of following instructions. I saved the Original one I was working with reading inbetween the line under file "day_2". not sure if the assignment was a simon says type of game so I followed as instructed after have finishing my project one. Either way "day2.html" would serve as  a saved template for future use for creating a template would just need to design with instyle css or external css for the form